RDA: Merge `ReachingDefinitionsAnalysis` and `DefUseAnalysis`
  * Current `DefUse` feels like "the wrong abstraction"
  * `RDA`, and `DefUse` have a lot in common: they just gather different
  data facts, but operate in a similar way
